Break `svm-state` into smaller modules
