+++ crates/state/Cargo.toml
@@ -16,6 +16,7 @@ blake3 = "1"
 futures = "0.3"
 sqlx = { version = "0.5", features = ["runtime-tokio-rustls", "sqlite", "macros"] }
 svm-hash = { path = "../hash" }
+svm-types = { path = "../types" }
 thiserror = "1"
 
 [dev-dependencies]
